float

+++ Planet Factory - Generator/Generateur_new/MapInfo.cpp
@@ -44,9 +44,9 @@ int MapInfo::editComponent(Component *c, int percent, e_etat etat)
     return 1;
 }
 
-int MapInfo::freeSpace(void)
+float MapInfo::freeSpace(void)
 {
-    int libre = 100;
+    float libre = 100;
     std::list<MyComponent*>::iterator it = _components.begin();
 
     for (it; it != _components.end(); it++)
